fix: add remaining refactor changes
