+++ tests/models/conftest.py
@@ -85,15 +85,15 @@ models = [
     ),
 ]
 if sys.version_info >= (3, 11):
-    from timecopilot.models.foundational.tirex import TiRex
+    from timecopilot.models.foundation.tirex import TiRex
 
     models.append(TiRex())
 
 if sys.version_info < (3, 13):
     from tabpfn_time_series import TabPFNMode
 
-    from timecopilot.models.foundational.sundial import Sundial
-    from timecopilot.models.foundational.tabpfn import TabPFN
+    from timecopilot.models.foundation.sundial import Sundial
+    from timecopilot.models.foundation.tabpfn import TabPFN
 
     models.append(TabPFN(mode=TabPFNMode.MOCK))
     models.append(Sundial(context_length=256, num_samples=10))
